Clear libxml errors before using them
In the FileImporter project I got failing hamcrest test due to totally
unrelated leftover libxml errors form other parts of the code. Clearing the
errors before listening and depending on them could be a solution to that.

Although this would not be needed if errors are cleared correctly in other code
places, it was really a pain to debug this.

See
https://gerrit.wikimedia.org/r/#/c/mediawiki/extensions/FileImporter/+/442895/14
https://integration.wikimedia.org/ci/job/quibble-vendor-mysql-hhvm-docker/7215/console
https://gerrit.wikimedia.org/r/#/c/mediawiki/extensions/FileImporter/+/445430/